Added most of next shuttles and part of live map.
For next shuttles, I implemented the nearby stops, favorite stops,
and all stops sections. Different from the last version, this
time I implemented next shuttles as a split view controller so
it will look better on the iPad. Nearby stops and favorite stops
work, and tapping on a stop brings up a stop view controller. On
the stop view controller, adding/removing the stop as a favorite
works, and the next departure times appear as they did in the old
verison of the app.

For the live map, I set up a map view using Apple MapKit and set
it to display the Stanford campus area when it loads. I also added
two basic buttons (just text for now) that zoom to campus and
the user's location.

One known bug I need to investigate is that, if a nearby stop is
close enough, the message that says "X feet" shows a really big
integer number that is definitely incorrect.